devotions/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
google-cloud-secret-manager==2.25.0
google-cloud-firestore==2.19.0
firebase-admin==6.2.0
pytz==2025.2
Authlib==1.3.1
Flask-Login==0.6.3
cryptography
requests
pandas
